requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
yt-dlp==2024.9.27
aiosqlite==0.20.0
python-multipart==0.0.12
pyyaml==6.0.2
aiofiles==24.1.0
